Added check that user own definition

+++ app/Http/Controllers/DefinitionController.php
@@ -113,10 +113,10 @@ class DefinitionController extends Controller
         // $out = new \Symfony\Component\Console\Output\ConsoleOutput();
         // $out->writeln('toot'.$definition->user()->first()['id']);
 
-        // if ($user->id != $definition->user()->first()['id']) {
-        //     return response()->json(null, 401);
-        // }
-        $definition->tags()->detach();
+        if ($user->id != $definition->user()->first()['id']) {
+            return response()->json(null, 401);
+        }
+        $definition->tags()->detach([]);
         $definition->delete();
 
         return response()->json(null, 204);
